Added `.all()` *blanket* leaf for *event observables* at the chain & pallet level
Right now, the type-hint given from `.all()` is slightly different from what is *actually* given from `.all`. This is because of limitations on TS's type-system when it comes to getting a list of properties given an object with properties.

At the moment, `.all()` gives a type-hint of `[<watch-leaf-1>] | [<watch-leaf-2>] | ...`, when In reality, it actually gives `[<watch-leaf-1>, <watch-leaf-2>, ...]`. I.e., instead of getting a list of length 1, which could be *any* of the potential watch leaves, what we *really* get is a list of many elements, where each one is a specific `WatchLeaf` from under the given *events* tree node